% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normalRAR}
\alias{normalRAR}
\title{Block Design for Response-Adaptive Randomization}
\usage{
normalRAR(mu_control, mu_treatment, sd_control, sd_treatment, N_total,
  block_number = 4, simulation = 10000, zvalue = NULL,
  rand_ratio = NULL, conf.int = 0.95, alternative = "less",
  replace = FALSE)
}
\arguments{
\item{mu_control}{scalar. Mean outcome in the control arm.}

\item{mu_treatment}{scalar. Mean outcome in the treatment arm.}

\item{sd_control}{scalar. Standard deviation of outcome in the control arm.}

\item{sd_treatment}{scalar. Standard deviation of outcome in the treatment}
}
\value{
a list with details on the simulation.
\describe{
  \item{\code{power}}{
    scalar. The power of the trial, ie. the proportion of success over the
    number of simulation ran.}
  \item{\code{N_enrolled}}{
    vector. The number of patients enrolled in the trial (sum of control
    and experimental group for each simulation. )}
  \item{\code{N_control}}{
    vector. The number of patients enrolled in the control group for
    each simulation.}
  \item{\code{N_control}}{
    vector. The number of patients enrolled in the experimental group for
    each simulation.}
}
}
\description{
Simulation for normal mean for block design for
   response-adaptive randomization with time as a confounding
}
\examples{
normalRAR(mu_control = 5.7, mu_treatment = 0.65,
          sd_control = 1.1, sd_treatment = 1.2,
          N_total = 200, block_number = 2, simulation = 100)

}
