% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_stratabayes.R
\name{prop_stratabayes}
\alias{prop_stratabayes}
\title{Stratified posterior mean for Binomial Data}
\usage{
prop_stratabayes(treatment, outcome, block, a0 = 0.5, b0 = 0.5,
  number_mcmc = 10000)
}
\arguments{
\item{treatment}{vector. The vector with treatment assignment, 0 for control
and 1 for treatment group.}

\item{outcome}{vector. The vector with outcome, 0 for failure
and 1 for success. Must be the same length as treatment variable.}

\item{block}{vector. The vector with factor level of the block.
Must be same lenhth as treatment variable.}

\item{a0}{scalar. Prior value for the beta rate \code{Beta(a0, b0)}.
Default is 0.5.}

\item{b0}{scalar. Prior value for the beta rate \code{Beta(a0, b0)}.
Default is 0.5.}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain draws in
sampling posterior.}
}
\value{
the weighted mean of posterior mean of difference (treatment - control).
}
\description{
Computing the posterior mean of treatment difference for
  stratified data with Bayesian method. The stratification is done over time.
}
\examples{
set.seed(20999)
prop_stratabayes(c(0, 1, 1, 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0),
                 c(0, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1),
                 as.factor(rep(1:3, each = 5)))
}
