% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomialRAR}
\alias{binomialRAR}
\title{Block Design for Response-Adaptive Randomization}
\usage{
binomialRAR(p_control, p_treatment, N_total, block_number = 4,
  simulation = 10000, zvalue = NULL, rand_ratio = NULL,
  conf.int = 0.95, alternative = "less", replace = FALSE)
}
\arguments{
\item{p_control}{scalar. Proportion of events under the control arm.}

\item{p_treatment}{scalar. Proportion of events under the treatment arm.}

\item{N_total}{scalar. Total sample size.}

\item{block_number}{scalar. Number of blocks or time levels. The default is set to 4.
If \code{block_number} is set to 1. This is a traditional RCT design.}

\item{simulation}{scalar. Number of simulation to be ran. The default is set to 10000.}

\item{zvalue}{vector. The z-value cutoff for corrected chi-square test statistics.}

\item{rand_ratio}{vector. The randomization ratio is set based on the corrected
chi-square test statistic. The length of rand_ratio should be the same length of
zvalue.}

\item{conf.int}{scalar. Confidence level of the interval.}

\item{alternative}{character. A string specifying the alternative hypothesis,
must be one of "less" (default) or "greater".}

\item{replace}{character. should sampling be with replacement? If replace is set to
FALSE (default), the 0 for control, 1 for treatment is replicated to the closest
integer and this vector is sampled with no replacement. If replace is set to TRUE,
the sampling is done based on randomization ratio provided with replacement.}
}
\value{
a list with power (prop of trial success), total sample size, sample size
   of patient in control and treatment group.
\describe{
  \item{\code{power}}{
    scalar. The power of the trial, ie. the proportion of success over the
    number of simulation ran.}
  \item{\code{N_enrolled}}{
    vector. The number of patients enrolled in the trial (sum of control
    and experimental group for each simulation. )}
  \item{\code{N_control}}{
    vector. The number of patients enrolled in the control group for
    each simulation.}
  \item{\code{N_control}}{
    vector. The number of patients enrolled in the experimental group for
    each simulation.}
}
}
\description{
Simulation for binomial counts for block design for
   response-adaptive randomization with time as a confounding
}
\examples{
binomialRAR(p_control = 0.7, p_treatment = 0.65, N_total = 200,
            block_number = 2, simulation = 100)
}
